+++ lib/src/argon2.dart
@@ -1,0 +1,5 @@
+// Copyright (c) 2023, Sudipto Chandra
+// All rights reserved. Check LICENSE file for details.
+
+export 'package:hashlib/src/algorithms/argon2.dart'
+    show Argon2Context, Argon2Type, Argon2Version;